# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/register": {
      "post": {
        "tags": [
          "Banking Application"
        ],
        "summary": "REGISTER Here",
        "description": "APIs create a Account",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "user register successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterRequest"
                }
              }
            }
          },
          "400": {
            "description": "user may already exist"
          }
        }
      }
    },
    "/api/logout": {
      "post": {
        "tags": [
          "Banking Application"
        ],
        "summary": "LogOut user",
        "description": "APIs for managing User logOut ",
        "operationId": "logout",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Not Found"
          },
          "200": {
            "description": "LogOut successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterRequest"
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": [
          "Banking Application"
        ],
        "summary": "Login user",
        "description": "APIs Login and stored database",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Login Successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterRequest"
                }
              }
            }
          },
          "400": {
            "description": "User Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "mobile": {
            "type": "string"
          }
        }
      },
      "UserLoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
    }
  }
}